cfg: Unify processing between C and Go
- Goto
- Labels
- Break
- Continue
