+++ src/control-flow/cfg-go.ts
@@ -1,16 +1,18 @@
 import type Parser from "web-tree-sitter";
 import type { BasicBlock, BuilderOptions, CFGBuilder } from "./cfg-defs";
 import {
+  getChildFieldText,
+  processBreakStatement,
+  processContinueStatement,
+  processGotoStatement,
+  processLabeledStatement,
+} from "./common-patterns.ts";
+import {
   type Context,
   GenericCFGBuilder,
   type StatementHandlers,
 } from "./generic-cfg-builder";
 import { type SwitchOptions, buildSwitch, collectCases } from "./switch-utils";
-
-function getChildFieldText(node: Parser.SyntaxNode, fieldName: string): string {
-  const child = node.childForFieldName(fieldName);
-  return child ? child.text : "";
-}
 
 const statementHandlers: StatementHandlers = {
   named: {
@@ -81,71 +83,6 @@ function defaultProcessStatement(
   );
   ctx.link.syntaxToNode(syntax, newNode);
   return { entry: newNode, exit: newNode };
-}
-
-function processGotoStatement(
-  gotoSyntax: Parser.SyntaxNode,
-  ctx: Context,
-): BasicBlock {
-  const name = gotoSyntax.firstNamedChild?.text as string;
-  const gotoNode = ctx.builder.addNode("GOTO", name, gotoSyntax.startIndex);
-  ctx.link.syntaxToNode(gotoSyntax, gotoNode);
-  return {
-    entry: gotoNode,
-    exit: null,
-    gotos: [{ node: gotoNode, label: name }],
-  };
-}
-function processLabeledStatement(
-  labelSyntax: Parser.SyntaxNode,
-  ctx: Context,
-): BasicBlock {
-  const name = getChildFieldText(labelSyntax, "label");
-  const labelNode = ctx.builder.addNode("LABEL", name, labelSyntax.startIndex);
-  ctx.link.syntaxToNode(labelSyntax, labelNode);
-  const labelContentSyntax = labelSyntax.namedChildren[1];
-  if (labelContentSyntax) {
-    const { entry: labeledEntry, exit: labeledExit } = ctx.state.update(
-      ctx.dispatch.single(labelContentSyntax),
-    );
-    if (labeledEntry) ctx.builder.addEdge(labelNode, labeledEntry);
-    return ctx.state.update({
-      entry: labelNode,
-      exit: labeledExit,
-      labels: new Map([[name, labelNode]]),
-    });
-  }
-  // Go allows for empty labels.
-  return ctx.state.update({
-    entry: labelNode,
-    exit: labelNode,
-    labels: new Map([[name, labelNode]]),
-  });
-}
-
-function processContinueStatement(
-  continueSyntax: Parser.SyntaxNode,
-  ctx: Context,
-): BasicBlock {
-  const continueNode = ctx.builder.addNode(
-    "CONTINUE",
-    "CONTINUE",
-    continueSyntax.startIndex,
-  );
-  ctx.link.syntaxToNode(continueSyntax, continueNode);
-  return { entry: continueNode, exit: null, continues: [continueNode] };
-}
-function processBreakStatement(
-  breakSyntax: Parser.SyntaxNode,
-  ctx: Context,
-): BasicBlock {
-  const breakNode = ctx.builder.addNode(
-    "BREAK",
-    "BREAK",
-    breakSyntax.startIndex,
-  );
-  ctx.link.syntaxToNode(breakSyntax, breakNode);
-  return { entry: breakNode, exit: null, breaks: [breakNode] };
 }
 
 function processForStatement(
